% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstepC}
\alias{EMstepC}
\title{Title here}
\usage{
EMstepC(W, C, P_Yosc, P_Xosc, B_T, Dat, sigX, sigY, sigH, sigT, sigTo, sigUo,
  invS = matrixXd::Zero(1, 1))
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
...
}
\description{
Title here
}

