% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstepC_fast}
\alias{EMstepC_fast}
\title{Title here}
\usage{
EMstepC_fast(W, C, B, X, Y, sigX, sigY, sigH, sigT, c1, c2, c3)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
...
}
\description{
Title here
}

