% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{Expect_M}
\alias{Expect_M}
\title{Performs the E step
Expectation step}
\usage{
Expect_M(X, Y, W, C, B, sigE, sigF, sigH, sigT, debug = F)
}
\arguments{
\item{X}{First dataset.}

\item{Y}{Second dataset.}

\item{W}{X loadings, p times r.}

\item{C}{Y loadings, q times r.}

\item{B}{Diagonal regression matrix with positive elements.}

\item{sigE}{Positive number. Standard deviation of noise in X}

\item{sigF}{Positive number. Standard deviation of noise in Y}

\item{sigH}{Positive number. Standard deviation of noise in latent space in Y}

\item{sigT}{Diagonal positive matrix. Standard deviations(!) of latent space in X}
}
\value{
List with expected first and second moments of the latent variables and noise.
No aggregation is done yet, this means that all Cxx are matrices.
This may be useful to check assumptions of no correlation between the latent variables.
}
\description{
Performs the E step
Expectation step
}

