% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglC}
\alias{loglC}
\title{Title here}
\usage{
loglC(W, C, P_Yosc, P_Xosc, B_T, Dat, sigX, sigY, sigH, sigT, sigTo, sigUo,
  c1 = 0, c2 = 0, c3 = 0)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
...
}
\description{
Title here
}

