% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglC_fast}
\alias{loglC_fast}
\title{Title here}
\usage{
loglC_fast(W, C, X, Y, sigX, sigY, sig2T, c1, c2, c3, Kc)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
...
}
\description{
Title here
}

