% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{Maximiz_M}
\alias{Maximiz_M}
\title{The M step}
\usage{
Maximiz_M(fit, X, Y, type = "SVD")
}
\arguments{
\item{fit}{A list as produced by \code{\link{Expect_M}}}

\item{X}{First dataset.}

\item{Y}{Second dataset.}
}
\value{
A list with updated estimates W, C, B, sigE, sigF, sigH and sigT.
}
\description{
Maximization step
}

