% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{meta_PPLSi}
\alias{meta_PPLSi}
\title{Performs PPLS fit in one direction. (use \link{PPLS} for fitting a PPLS model)}
\usage{
meta_PPLSi(X, Y, Ipopu, EMsteps = 100, atol = 1e-04,
  initialGuess = c("equal", "o2m", "random", "custom"), customGuess = NULL,
  critfunc = function(x) {     x }, constraints = fconstraint())
}
\arguments{
\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}

\item{EMsteps}{strictly positive integer. Denotes the maximum number of EM steps to make.}

\item{atol}{Double, convergence criterium for the log-likelihood}

\item{initialGuess}{A string. Choose "o2m", "random" or "equal" depending on what type of initial guess you want.}

\item{customGuess}{A list with named components W,C,B,sigE,sigF,sigH,sigT.}

\item{critfunc}{Function measuring the increment value, i.e. f(L_{i+1} - L_{i}). Usually f == I or f == abs.}

\item{constraints}{List. Ideally being constructed with \link{fconstraint}.}
}
\value{
A list with estimates for\itemize{
   \item{W}{Matrix}
   \item{C}{Matrix}
   \item{B}{Matrix}
   \item{sig}{ Vector containing updated sigX, sigY, sigH and sigT}
   \item{logvalue}{ Vector with last log-likelihood(s)}
   \item{Last_increment}{ Double, value of last increment.}
   \item{Number_steps}{ Integer, number of steps needed to stop.}
 }
}
\description{
Performs PPLS fit in one direction. (use \link{PPLS} for fitting a PPLS model)
}
\details{
This function estimates loadings and variances for one direction.
}

