\name{normalize.vector}
\alias{normalize.vector}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Normalization of a vector}
\description{
  normalizes a vector to unit length
}
\usage{
normalize.vector(v)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{v}{vector}
}

\value{normalized vector 
}
\author{Nicole Kraemer}
\note{This is an auxiliary function.
}
\examples{
v<-1:5
w<-normalize.vector(v)
}
\keyword{math}
