% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{plot.PPLS}
\alias{plot.PPLS}
\title{Plot function for class PPLS.}
\usage{
\method{plot}{PPLS}(x, XorY = c("X", "Y"), i = 1, j = NULL,
  use_ggplot2 = TRUE, ...)
}
\arguments{
\item{x}{A PPLS fit (an object of class PPLS)}

\item{XorY}{Plot loadings for X or Y?}

\item{i}{Positive integer. The i-th loading/score will be plotted on the first axis}

\item{j}{Positive integer. The j-th loading/score will be plotted on the second axis}

\item{use_ggplot2}{Logical. Use ggplot2?}

\item{...}{For consistency}
}
\value{
If ggplot2 is active, returns the ggplot object. Else NULL.
}
\description{
Plot function for class PPLS.
}
\details{
This function plots two loadings/scores.
}

