% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{PPLS_simult}
\alias{PPLS_simult}
\title{The PPLS fitting function}
\usage{
PPLS_simult(X, Y, a, EMsteps = 10, atol = 1e-04, type = c("SVD", "QR"),
  lasso_lambda = c(0, 0), ...)
}
\arguments{
\item{X}{First dataset.}

\item{Y}{Second dataset.}

\item{a}{Positive integer, number of components}

\item{EMsteps}{Positive integer, number of EM steps}

\item{atol}{positive double, convergence criterium}

\item{type}{String. One of "SVD" or "QR"}

\item{...}{debug, logical. Should the slow Expect_M be used?}
}
\value{
list of class PPLS_simul with expectations, loglikelihoods and estimates.
}
\description{
Simultaneous PPLS fitting function
}
