% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{PPLSi}
\alias{PPLSi}
\title{Performs PPLS fit in one direction.}
\usage{
PPLSi(X, Y, EMsteps = 100, atol = 1e-04, initialGuess = c("o2m", "random",
  "equal"), critfunc = function(x) {     x }, print_logvalue = FALSE)
}
\arguments{
\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}

\item{EMsteps}{strictly positive integer. Denotes the maximum number of EM steps to make.}

\item{atol}{Double, convergence criterium for the log-likelihood}

\item{initialGuess}{A string. Choose "o2m", "random" or "equal" depending on what type of initial guess you want.}

\item{critfunc}{Function measuring the increment value, i.e. f(L_{i+1} - L_{i}). Usually f == I or f == abs.}

\item{print_logvalue}{Logical. Should we print log-likelihood values for each step?}
}
\value{
A list with estimates for\itemize{
   \item{W}{Matrix}
   \item{C}{Matrix}
   \item{B}{Matrix}
   \item{sig}{Vector containing updated sigX, sigY, sigH and sigT}
   \item{logvalue}{Vector with last (or all, if print_logvalue==TRUE) log-likelihood(s)}
   \item{Last_increment}{Double, value of last increment.}
   \item{Number_steps}{Integer, number of steps needed to stop.}
 }
}
\description{
Performs PPLS fit in one direction.
}
\details{
This function estimates loadings and variances for one direction.
}

