% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval_PPLS.R
\name{predict.PPLS}
\alias{predict.PPLS}
\title{Predicts X or Y with PPLS}
\usage{
\method{predict}{PPLS}(object, newdata, XorY = c("X", "Y"), ...)
}
\arguments{
\item{object}{PPLS fit}

\item{newdata}{New data, which one of X or Y is specified in \code{XorY}.}

\item{XorY}{Character specifying whether \code{newdata} is X or Y.}
}
\value{
Predicted Data
}
\description{
Predicts X or Y based on new data on Y or X
}
