% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{print.PPLS}
\alias{print.PPLS}
\title{Print function for class PPLS.}
\usage{
\method{print}{PPLS}(x, perc = FALSE, ...)
}
\arguments{
\item{x}{A PPLS fit (an object of class PPLS)}

\item{perc}{Display relative percentages yes/no?}

\item{...}{For consistency}
}
\value{
Invisible, the outputted table
}
\description{
Print function for class PPLS.
}
\details{
This function shows the absolute/relative variances of each component.
}

