% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\docType{package}
\name{Probabilistic-PLS}
\alias{Probabilistic-PLS}
\alias{Probabilistic-PLS-package}
\title{PPLS: Probabilistic Partial Least Squares}
\description{
This package implements the Partial Least Squares method in a probabilistic framework.
}
\section{Usage}{

Just do PPLS(Datamatrix_1,Datamatrix_2,nr_of_components) to have a quick and dirty fit, modify the default arguments if necessary.
}

\author{
Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}),
Jeanine Houwing-Duistermaat (\email{J.J.Houwing@lumc.nl}),
Geurt Jongbloed (\email{G.Jongbloed@tudelft.nl}),
Szymon Kielbasa (\email{S.M.Kielbasa@lumc.nl}),
Hae-Won Uh (\email{H.Uh@lumc.nl}).

Maintainer: Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}).
}
\keyword{Probabilistic-PLS}
