% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{scores.PPLS}
\alias{scores.PPLS}
\title{Get scores from PPLS fit.}
\usage{
scores.PPLS(fit, X, Y, subset = NULL)
}
\arguments{
\item{fit}{A PPLS fit (an object of class PPLS)}

\item{X}{Matrix}

\item{Y}{Matrix}

\item{subset}{vector of positive integers denoting with components you want.}
}
\value{
Both the X and Y scores concatenated.
}
\description{
Get scores from PPLS fit.
}
\details{
This function plots two loadings/scores.
}

