% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulC}
\alias{simulC}
\title{Title here}
\usage{
simulC(N, W, C, P_Yosc, P_Xosc, B_T, sigX, sigY, sigH, sigT, sigTo, sigUo)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
...
}
\description{
Title here
}

