% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{variances.PPLS_simult}
\alias{variances.PPLS_simult}
\title{Variances for PPLS_simult}
\usage{
variances.PPLS_simult(fit, data, XorY = c("X", "Y"))
}
\arguments{
\item{fit}{A list as produced by \code{\link{Expect_M}}}

\item{data}{Data matrix, X or Y}

\item{XorY}{Which data matrix did you supply, X or Y?}
}
\value{
SE's for the loadings of corresponding data (so W or C)
}
\description{
Calculates asymptotic variances for PPLS loadings
}

