% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{EMstep_W}
\alias{EMstep_W}
\title{Performs one EM step}
\usage{
EMstep_W(X, Y, W. = W, C. = C, B_T. = B_T, sigX. = sigX, sigY. = sigY,
  sigH. = sigH, sigT. = sigT)
}
\arguments{
\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}

\item{W.}{Numeric matrix.}

\item{C.}{Numeric matrix.}

\item{B_T.}{Numeric}

\item{sigX.}{Numeric}

\item{sigY.}{Numeric}

\item{sigH.}{Numeric}

\item{sigT.}{Numeric}
}
\value{
A list with updated values for\itemize{
   \item{W}{Matrix}
   \item{C}{Matrix}
   \item{B}{Matrix}
   \item{sighat}{Vector containing updated sigX and sigY}
   \item{siglathat}{Vector containing sigH and sigT}
 }
}
\description{
Performs one EM step
}
\details{
This function passes its arguments to EMstepC (a C++ function, see \link{Rcpp}), which returns the expected sufficient statistics.
The maximization is done afterwards in this function. This may become a full C++ function later.
}

