% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\name{logl_W}
\alias{logl_W}
\title{Calculates loglikelihood for PPLS model.}
\usage{
logl_W(X, Y, W., C., B_T., sigX., sigY., sigH., sigT.)
}
\arguments{
\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}

\item{W.}{Numeric matrix.}

\item{C.}{Numeric matrix.}

\item{B_T.}{Numeric matrix.}

\item{sigX.}{Numeric matrix.}

\item{sigY.}{Numeric matrix.}

\item{sigH.}{Numeric matrix.}

\item{sigT.}{Numeric matrix.}
}
\value{
The log-likelihood (i.e. scalar) of the parameters given the observed data.
}
\description{
Calculates loglikelihood for PPLS model.
}
\details{
This function passes its arguments to LoglC_fast (a C++ function, see \link{Rcpp}).
}
