% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EM_W_multi.R
\docType{package}
\name{PPLS}
\alias{PPLS}
\alias{PPLS-package}
\title{PPLS: Probabilistic Partial Least Squares}
\usage{
PPLS(X, Y, nr_comp = 1, EMsteps = 100, atol = 1e-04,
  initialGuess = c("equal", "o2m", "random"), critfunc = function(x) {     x
  }, print_logvalue = FALSE)
}
\arguments{
\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}

\item{nr_comp}{Strictly positive integer. The number of PPLS components to fit.}

\item{EMsteps}{strictly positive integer. Denotes the maximum number of EM steps to make.}

\item{atol}{Double, convergence criterium for the log-likelihood}

\item{initialGuess}{A string. Choose "o2m", "random" or "equal" depending on what type of initial guess you want.}

\item{critfunc}{Function measuring the increment value, i.e. f(L_{i+1} - L_{i}). Usually f == I or f == abs.}

\item{print_logvalue}{Logical. Should we print log-likelihood values for each step?}
}
\value{
A list with estimates for\itemize{
   \item{W}{Matrix, each column is a fit.}
   \item{C}{Matrix, each column is a fit.}
   \item{B}{Matrix, each column is a fit.}
   \item{sig}{Matrix containing updated sigX, sigY, sigH and sigT. Columns are named.}
   \item{Other_output}{\itemize{
         \item{Last_increment}{.}
         \item{Number_steps}{.}
         \item{Loglikelihoods}{.}
         }
   }
 }
}
\description{
This package implements the Partial Least Squares method in a probabilistic framework.

Performs PPLS fit in sequentially multiple directions.
}
\details{
This function estimates loadings and variances for multiple orthogonal directions.
}
\section{Usage}{

Just do PPLS(Datamatrix_1,Datamatrix_2,nr_of_components) to have a quick and dirty fit, modify the default arguments if necessary.
}
\author{
Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}),
Jeanine Houwing-Duistermaat (\email{J.J.Houwing@lumc.nl}),
Geurt Jongbloed (\email{G.Jongbloed@tudelft.nl}),
Szymon Kielbasa (\email{S.M.Kielbasa@lumc.nl}),
Hae-Won Uh (\email{H.Uh@lumc.nl}).

Maintainer: Said el Bouhaddani (\email{s.el_bouhaddani@lumc.nl}).
}
\keyword{Probabilistic-PLS}

